Clean up set seed and joint angle targest
