+++ grasp_generation/scripts/validate_grasps.py
@@ -25,38 +25,7 @@ from utils.hand_model_type import (
 from utils.qpos_pose_conversion import qpos_to_pose, qpos_to_translation_rot_jointangles
 from typing import List
 import math
-import random
-
-
-def set_seed(seed, torch_deterministic=False, rank=0):
-    """set seed across modules"""
-    if seed == -1 and torch_deterministic:
-        seed = 42 + rank
-    elif seed == -1:
-        seed = np.random.randint(0, 10000)
-    else:
-        seed = seed + rank
-
-    print("Setting seed: {}".format(seed))
-
-    random.seed(seed)
-    np.random.seed(seed)
-    torch.manual_seed(seed)
-    os.environ["PYTHONHASHSEED"] = str(seed)
-    torch.cuda.manual_seed(seed)
-    torch.cuda.manual_seed_all(seed)
-
-    if torch_deterministic:
-        # refer to https://docs.nvidia.com/cuda/cublas/index.html#cublasApi_reproducibility
-        os.environ["CUBLAS_WORKSPACE_CONFIG"] = ":4096:8"
-        torch.backends.cudnn.benchmark = False
-        torch.backends.cudnn.deterministic = True
-        torch.use_deterministic_algorithms(True)
-    else:
-        torch.backends.cudnn.benchmark = True
-        torch.backends.cudnn.deterministic = False
-
-    return seed
+from utils.seed import set_seed
 
 
 def compute_loss(
@@ -101,13 +70,13 @@ def compute_loss(
         nearest_normals = torch.gather(
             normals, 1, nearest_point_index.reshape(-1, 1, 1).expand(-1, 1, 3)
         )
-        admited = -nearest_distances < dist_thresh_to_move_finger
-        admited = admited.reshape(-1, 1, 1).expand(-1, 1, 3)
+        admitted = -nearest_distances < dist_thresh_to_move_finger
+        admitted = admitted.reshape(-1, 1, 1).expand(-1, 1, 3)
         contact_points_hand[:, i : i + 1, :] = torch.where(
-            admited, nearest_points_hand, contact_points_hand[:, i : i + 1, :]
+            admitted, nearest_points_hand, contact_points_hand[:, i : i + 1, :]
         )
         contact_normals[:, i : i + 1, :] = torch.where(
-            admited, nearest_normals, contact_normals[:, i : i + 1, :]
+            admitted, nearest_normals, contact_normals[:, i : i + 1, :]
         )
 
     target_points = contact_points_hand + contact_normals * dist_move_link
